Rename and reorganize packages
