+++ src/main/java/me/often/aureliummobs/listeners/PlayerJoinLeave.java
@@ -1,6 +1,6 @@
-package me.often.aureliummobs.Listeners;
+package me.often.aureliummobs.listeners;
 
-import me.often.aureliummobs.Main.Main;
+import me.often.aureliummobs.Main;
 import org.bukkit.event.EventHandler;
 import org.bukkit.event.Listener;
 import org.bukkit.event.player.PlayerJoinEvent;
