+++ src/main/resources/plugin.yml
@@ -1,4 +1,4 @@
-main: me.often.aureliummobs.Main.Main
+main: me.often.aureliummobs.Main
 name: AureliumMobs
 author: _OfTeN_
 version: 1.14
